requirements.txt
Django
Pillow
django-shortuuidfield
os